move some function into different JS file
